chore: updates after review
